IMPRO1039 Temporal interpolation of UV data - part 1
