+++ tests/improver-temporal-interpolate/00-null.bats
@@ -36,6 +36,7 @@
 usage: improver-temporal-interpolate [-h] [--profile]
                                      [--profile_file PROFILE_FILE]
                                      (--interval_in_mins INTERVAL_IN_MINS | --times TIMES [TIMES ...])
+                                     [--interpolation_method INTERPOLATION_METHOD]
                                      --output_files OUTPUT_FILES
                                      [OUTPUT_FILES ...]
                                      INFILES INFILES
